utilities module to share placeCursorAtEnd function
